@@ -1,5 +1,5 @@
 {
-  "name": "autocad-to-excel-api",
+  "name": "autocad-vue3",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,11 +8,10 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@popperjs/core": "^2.11.6",
     "bootstrap": "^5.2.2",
     "core-js": "^3.8.3",
-    "vue": "^2.6.14",
-    "vue-router": "^3.5.1",
+    "vue": "^3.2.13",
+    "vue-router": "^4.0.3",
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
@@ -28,7 +27,6 @@
     "eslint-plugin-vue": "^8.0.3",
     "prettier": "^2.4.1",
     "sass": "^1.32.7",
-    "sass-loader": "^12.0.0",
-    "vue-template-compiler": "^2.6.14"
+    "sass-loader": "^12.0.0"
   }
 }
